requirements.txt
fastapi==0.121.2
uvicorn==0.34.0
pydantic==2.10.5
starlette==0.49.3
numpy==1.26.4
google-generativeai==0.8.5
python-dotenv==1.0.0
requests==2.31.0
